#385-add-use-enroll-course-hook
