+++ src/store/courses/types.ts
@@ -2,10 +2,8 @@ import type { CoursesModel } from 'api/courses/types';
 import { ProcessEnum } from 'utils/constants';
 
 export type EnrollStatusType = {
-  [key: string]: {
-    process: ProcessEnum;
-    error: string | null;
-  };
+  process: ProcessEnum;
+  error: string | null;
 };
 
 export type CoursesThunkState = {
@@ -13,7 +11,7 @@ export type CoursesThunkState = {
   error: string | null;
   count: CoursesModel['count'];
   courses: CoursesModel['results'];
-  enrollStatus: EnrollStatusType;
+  enrollStatus: Record<string, EnrollStatusType>;
 };
 
 export type CoursesState = CoursesThunkState;
